fixed variable on ft_strlen

+++ ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: gdelarme <gdelarme@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/10 18:39:06 by gdelarme          #+#    #+#             */
-/*   Updated: 2021/11/18 18:27:23 by gdelarme         ###   ########.fr       */
+/*   Updated: 2021/11/22 21:06:58 by gdelarme         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *s)
+size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != '\0')
